# openapi/0.1.0.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SPPA Core HTTP-artifact binding",
    "version": "0.1.0",
    "description": "Working draft. Provider binding, not Hub search API. Production requires HTTPS; loopback reference is a development harness.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "https://provider.example",
      "description": "Replace with the trusted provider origin"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
...
    "/v1/jobs": {
      "post": {
        "operationId": "job_create",
        "summary": "job.create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "type": {
                        "const": "job.create"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful job.create",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "const": "job.created"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Structured SPPA error; see error-code HTTP mapping",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "const": "error"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "200": {
            "description": "Completed synchronous capability, if advertised",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "const": "job.created"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/get": {
      "post": {
        "operationId": "job_get",
        "summary": "job.get",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "type": {
                        "const": "job.get"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful job.get",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "const": "job.snapshot"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Structured SPPA error; see error-code HTTP mapping",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "const": "error"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/cancel": {
      "post": {
        "operationId": "job_cancel",
        "summary": "job.cancel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "type": {
                        "const": "job.cancel"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful job.cancel",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "const": "job.snapshot"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Structured SPPA error; see error-code HTTP mapping",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "const": "error"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/retry": {
      "post": {
        "operationId": "job_retry",
        "summary": "job.retry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "type": {
                        "const": "job.retry"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful job.retry",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "const": "job.created"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Structured SPPA error; see error-code HTTP mapping",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "const": "error"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/artifacts/get": {
      "post": {
        "operationId": "artifact_get",
        "summary": "artifact.get",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "type": {
                        "const": "artifact.get"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful artifact.get",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "const": "artifact.metadata"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Structured SPPA error; see error-code HTTP mapping",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "https://theparitt.github.io/sppa-protocol/schemas/0.1.0/envelope.schema.json"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "const": "error"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}